Copyright component abstracted.  Code clean up done.
