+++ src/components/App/App.js
@@ -1,10 +1,9 @@
 import { useEffect, useState, useContext } from 'react';
-import { BrowserRouter, Routes, Route } from 'react-router-dom';
+import { Routes, Route } from 'react-router-dom';
 import './App.css';
 import { CssBaseline, Container } from '@mui/material';
 import NavBar from '../NavBar/NavBar';
 import Home from '../Home/Home';
-import Category from '../Category/Category';
 import SignIn from '../SignIn/SignIn';
 import SignUp from '../SignUp/SignUp';
 import NotFound from '../NotFound/NotFound';
@@ -16,44 +15,25 @@ import Wishlist from '../Wishlist/Wishlist.js';
 import Cart from '../Cart/Cart';
 import ProtectedRoute from '../ProtectedRoute/ProtectedRoute';
 import StickyFooter from '../StickyFooter/StickyFooter';
-// import { AuthProvider } from '../../contexts/AuthContext'
-import {isCartItemFound, getCartItem, updateCartItem, addToExistingCartItem, addNewCartItem, removeFromExistingCartItem, clearCartItem, isFaveItem, getFaveItem, addFaveItem, removeFaveItem, calculateCartCount} from '../../utils/utilities'
-// import data from '../../data/data.json'
-
-const initialUser = {
-  displayName: 'Test User',
-  email: 'test@test.com',
-  favourites: [],
-}
+import {isCartItemFound, getCartItem, updateCartItem, addToExistingCartItem, addNewCartItem, removeFromExistingCartItem, clearCartItem, isFaveItem, addFaveItem, removeFaveItem, calculateCartCount} from '../../utils/utilities'
 
 function App() {
-  const [isLoading, setIsLoading] = useState(true) //custom hook for isLoading
+  const [isLoading, setIsLoading] = useState(true)
   const [data, setData] = useState([])
   const [error, setError] = useState('')
-  // const [isSignedIn, setIsSignedIn] = useState(false) //move to AuthContext
-  const [cartItems, setCartItems] = useState([]) //move to UserContext
+  const [cartItems, setCartItems] = useState([]) 
   const [cartCount, setCartCount] = useState(0)
   const [favourites, setFavourites] = useState([])
-  // const [user, setUser] = useState(initialUser) //get displayName, email, favourites, cartItems in UserContext
-  const getUrl = () => {
-    // if (process.env.NODE_ENV !== 'production') {
-    //   return '/data/data.json'
-    // }
-    return '/.netlify/functions/getShopData'
-  } 
 
   const getItemFromInventory = (itemId) => {
     for(const category in data.products) {
       let foundItem = data.products[category].find(item => item.id === itemId)
-      console.log(foundItem, {...foundItem, ...{category}})
       if(foundItem) return {...foundItem, ...{category}}
     }
   }
 
   const updateCartCount = (cartItems) => {
-    console.log('updateCartCount cartItems', cartItems)
     const newCartCount = calculateCartCount(cartItems)
-    console.log(newCartCount)
     setCartCount(newCartCount)
   }
 
@@ -68,7 +48,6 @@ function App() {
         newCartItems = addToExistingCartItem(cartItems, id)
       } else {
         return alert('Not enough stock')
-        // return { error: 'Not enough stock'} //toastify it?
       }
     }
     setCartItems(newCartItems)
@@ -76,28 +55,19 @@ function App() {
   }
   //called by minus '-' or '<' button from shopping cart
   const removeItemFromCart = (id) => {
-    // const cartItemIndex = isCartItemFound(cartItems, id)
-    console.log('removeItemFromCart', id)
-    console.log(isCartItemFound(cartItems, id))
     if(!isCartItemFound(cartItems, id)) return
     const cartItemToRemove = getCartItem(cartItems, id)
-    console.log(cartItemToRemove)
     let newCartItems
-    if(cartItemToRemove.quantity <= 1 && cartItemToRemove.quantity > 0) {
+    if(cartItemToRemove.quantity <= 1 && cartItemToRemove.quantity > 0) { //1 left
       newCartItems = clearCartItem(cartItems, id)
-      console.log('1 left')
     } else {
-      newCartItems = removeFromExistingCartItem(cartItems, id)
-      console.log('at least 2 left')
+      newCartItems = removeFromExistingCartItem(cartItems, id) //at least 2 left
     }
-    console.log(newCartItems)
     setCartItems(newCartItems)
     updateCartCount(newCartItems)
   }
   //called by delete 'x' button from shopping cart
   const clearItemFromCart = (id) => {
-    // const cartItemIndex = isCartItemFound(cartItems, id)
-    // if(cartItemIndex < 0) return
     if(!isCartItemFound(cartItems, id)) return
     const newCartItems = clearCartItem(cartItems, id)
     setCartItems(newCartItems)
@@ -105,31 +75,16 @@ function App() {
   }
 
   const setNewQuantityForCartItem = (id, quantity) => {
-    // let newCartItems
-    // if(!isCartItemFound(cartItems, id)) {
-    //   const newCartItem = getItemFromInventory(id)
-    //   newCartItems = addNewCartItem(cartItems, newCartItem)
-    // } else {
-    //   newCartItems = addToExistingCartItem(cartItems, id)
-    // }
-    // setCartItems(newCartItems)
-    // updateCartCount(newCartItems)
-    console.log('setNewQuantityForCartItem: id, quantity', {id, quantity})
     if(quantity === 0) clearItemFromCart(id)
-    console.log(isCartItemFound(cartItems, id))
-    // if(!isCartItemFound(cartItems, id)) return
+    
     let newCartItems
     if(!isCartItemFound(cartItems, id)) {
       const newCartItem = getItemFromInventory(id)
-      console.log(newCartItem, quantity)
       newCartItems = addNewCartItem(cartItems, {...newCartItem, ...{ quantity }})
-      console.log(newCartItems, quantity)
 
     } else {
       const newCartItem = getCartItem(cartItems, id)
-      console.log(newCartItem)
       newCartItems = updateCartItem(cartItems, {...newCartItem, ...{quantity}})
-      console.log(newCartItems)
     }
     setCartItems(newCartItems)
     updateCartCount(newCartItems)
@@ -139,39 +94,21 @@ function App() {
     return isFaveItem(favourites, id)
   }
   const addToFavourites = (id) => {
-    // console.log(e.target)
-    console.log('fav add',id)
     if(isFaveItem(favourites, id)) return
     const newFavourites = addFaveItem(favourites, id)
     setFavourites(newFavourites)
   }
 
   const removeFromFavourites = (id) => {
-    console.log('fav remove', id)
     if(!isFaveItem(favourites, id)) return
     const newFavourites = removeFaveItem(favourites, id)
     setFavourites(newFavourites)
   }
-
-  
-  // const signIn = (signinData) => {
-  //   const {email, password} = signinData
-  //   //pull data from user authentication and firestore
-  //   console.log('in App signIn fn', {email, password})
-  //   const newUser = {
-  //     displayName: 'Annie Smith',
-  //     email,
-  //     favourites: [],
-  //   }
-  //   setUser(newUser)
-  //   setIsSignedIn(true)
-  // }
   
   useEffect(() => {
-      // setIsLoading(true)
       const fetchData = async () => {
           try {
-            const url = getUrl()
+            const url = '/.netlify/functions/getShopData'
             const response = await fetch(url)
           if (!response.ok || response.status !== 200 || response.status >= 400) throw new Error('Unable to fetch data')
             const data = await response.json()
@@ -183,7 +120,6 @@ function App() {
           }
       }
       fetchData()
-      // console.log(data)
       // eslint-disable-next-line react-hooks/exhaustive-deps
       }, [])
