+++ src/components/Review/Review.js
@@ -7,45 +7,10 @@ import ListItemText from '@mui/material/ListItemText';
 import Grid from '@mui/material/Grid';
 import { parseIntToDollarsAndCents, calculateCartSubtotal } from '../../utils/utilities'
 
-const products = [
-  {
-    name: 'Product 1',
-    desc: 'A nice thing',
-    price: '$9.99',
-  },
-  {
-    name: 'Product 2',
-    desc: 'Another thing',
-    price: '$3.45',
-  },
-  {
-    name: 'Product 3',
-    desc: 'Something else',
-    price: '$6.51',
-  },
-  {
-    name: 'Product 4',
-    desc: 'Best thing of all',
-    price: '$14.11',
-  },
-  { name: 'Shipping', desc: '', price: 'Free' },
-];
-
-const addresses = ['1 MUI Drive', 'Reactville', 'Anytown', '99999', 'USA'];
-const payments = [
-  { name: 'Card type', detail: 'Visa' },
-  { name: 'Card holder', detail: 'Mr John Smith' },
-  { name: 'Card number', detail: 'xxxx-xxxx-xxxx-1234' },
-  { name: 'Expiry date', detail: '04/2024' },
-];
-
 const addressShape = ['addressLine1', 'addressLine2', 'cityTownVillage', 'stateProvinceRegion', 'postCode', 'country']
 
-export default function Review({cartItems, addressFormDataForShip, addressFormDataForBill, isShippingEqualBillingAddress, orderSummary, setOrderSummary}) {
+export default function Review({cartItems, addressFormDataForShip, addressFormDataForBill}) {
   const [subtotal, setSubtotal] = React.useState(0)
-  // const stripe = useStripe()
-  // const elements = useElements()
-  console.log(addressFormDataForShip)
   React.useEffect(() => {
     const newSubtotal = calculateCartSubtotal(cartItems)
     setSubtotal(newSubtotal)
@@ -69,24 +34,12 @@ export default function Review({cartItems, addressFormDataForShip, addressFormDa
               )
             })
         }
-        {/* {products.map((product) => (
-          <ListItem key={product.name} sx={{ py: 1, px: 0 }}>
-            <ListItemText primary={product.name} secondary={product.desc} />
-            <Typography variant="body2">{product.price}</Typography>
-          </ListItem>
-        ))} */}
         <ListItem sx={{ py: 1, px: 0 }}>
           <ListItemText primary="Total" />
           <Typography variant="subtitle1" sx={{ fontWeight: 700 }}>
             AUD {parseIntToDollarsAndCents(subtotal)}
           </Typography>
         </ListItem>
-        {/* <ListItem sx={{ py: 1, px: 0 }}>
-          <ListItemText primary="Total" />
-          <Typography variant="subtitle1" sx={{ fontWeight: 700 }}>
-            $34.06
-          </Typography>
-        </ListItem> */}
       </List>
       <Grid container spacing={2}>
         <Grid item xs={12} sm={6}>
@@ -124,18 +77,6 @@ export default function Review({cartItems, addressFormDataForShip, addressFormDa
             Payment 
           </Typography>
           <CardElement />
-          {/* <Grid container>
-            {payments.map((payment) => (
-              <React.Fragment key={payment.name}>
-                <Grid item xs={6}>
-                  <Typography gutterBottom>{payment.name}</Typography>
-                </Grid>
-                <Grid item xs={6}>
-                  <Typography gutterBottom>{payment.detail}</Typography>
-                </Grid>
-              </React.Fragment>
-            ))}
-          </Grid> */}
         </Grid>
       </Grid>
     </React.Fragment>
