+++ src/components/SubCategory/SubCategory.js
@@ -1,19 +1,26 @@
 import {useParams, useNavigate} from 'react-router-dom'
-import { Stack, Container, IconButton, Grid, Card, CardActionArea, CardMedia, CardContent, Typography, ImageList, ImageListItem, ImageListItemBar, ListSubheader } from '@mui/material'
+// import { Stack, Container, IconButton, Grid, Card, CardActionArea, CardMedia, CardContent, Typography } from '@mui/material'
+import Stack from '@mui/material/Stack'
+import Container from '@mui/material/Container'
+import IconButton from '@mui/material/IconButton'
+import Grid from '@mui/material/Grid'
+import Card from '@mui/material/Card'
+import CardActionArea from '@mui/material/CardActionArea'
+import CardMedia from '@mui/material/CardMedia'
+import CardContent from '@mui/material/CardContent'
+import Typography from '@mui/material/Typography'
 import FavoriteIcon from '@mui/icons-material/Favorite'
 import AddShoppingCartIcon from '@mui/icons-material/AddShoppingCart';
 
 const SubCategory = ({data, favourites, addItemToCart, isFaveFound, addToFavourites, removeFromFavourites}) => {
     const {category, subcategory} = useParams()
     const navigate = useNavigate()
-    console.log(favourites)
     
     const handleClickNavigate = (e) => {
         navigate(`/${category}/${subcategory}/${e.target.id}`)
     }
 
     const handleClickCart = (id) => {
-        console.log('added to cart', id)
         addItemToCart(id)
     }
 
@@ -33,23 +40,18 @@ const SubCategory = ({data, favourites, addItemToCart, isFaveFound, addToFavouri
                     .filter(item => item.subcategory === subcategory)
                     .map(item => {
                         const {id, name, description, image, amount, currency} = item 
-                        // console.log('item', item)
                         return (
                             <Grid item xs={6} sm={4} md={3} key={`${name}-${id}`}>
                                 <Card sx={{ maxWidth: 345, height: 320, overflowY: 'scroll', position: 'relative' }}>
                                     <CardContent>
-                                        {/* <Typography gutterBottom variant="h8" component="div" sx={{position:'absolute', top: 0, left:0, right:0, marginRight: 0, marginLeft: 'auto', zIndex: 10, background: 'rgba(255,255,255,0)'}}> */}
-                                        {/* <Stack direction="row" sx={{position:'absolute', top: 0, left:0, right:0, marginRight: 0, marginLeft: 'auto', zIndex: 10, background: 'rgba(255,255,255,0)'}}> */}
                                         <Stack direction="row-reverse" sx={{background: 'rgba(255,255,255,0)'}}>
                                             <IconButton onClick={() => handleClickCart(id)} aria-label={`add product id ${id} to cart`}>
                                                 <AddShoppingCartIcon id={`cart-${id}`} role='button' aria-label={`add product id ${id} to cart`} />
                                             </IconButton>
                                             <IconButton onClick={() => handleClickFave(id)} aria-label={`add product id ${id} to wishlist`}>
-                                                {/* <FavoriteIcon id={`fav-${id}`} role='button' sx={{ color: 'rgb(255,20,147)'}} aria-label={`add product id ${id} to wishlist`}/> */}
                                                 <FavoriteIcon id={`fav-${id}`} role='button' sx={{ color: isFaveFound(id) ? 'rgb(255,20,147)' : 'rgb(0,0,0, 0.35)'}} aria-label={`add product id ${id} to wishlist`}/>
                                             </IconButton>
                                         </Stack>
-                                        {/* </Typography> */}
                                     </CardContent>
                                     <CardActionArea onClick={handleClickNavigate}>
                                         <CardMedia
