+++ src/components/Wishlist/Wishlist.js
@@ -1,6 +1,15 @@
 import { useState, useEffect, useContext } from 'react'
 import { useNavigate } from 'react-router-dom'
-import { Container, Typography, Grid, Card, CardActionArea, CardMedia, CardContent, Stack, IconButton } from '@mui/material'
+// import { Container, Typography, Grid, Card, CardActionArea, CardMedia, CardContent, Stack, IconButton } from '@mui/material'
+import Container from '@mui/material/Container'
+import Typography from '@mui/material/Typography'
+import Grid from '@mui/material/Grid'
+import Card from '@mui/material/Card'
+import CardActionArea from '@mui/material/CardActionArea'
+import CardMedia from '@mui/material/CardMedia'
+import CardContent from '@mui/material/CardContent'
+import Stack from '@mui/material/Stack'
+import IconButton from '@mui/material/IconButton'
 import FavoriteIcon from '@mui/icons-material/Favorite'
 import AddShoppingCartIcon from '@mui/icons-material/AddShoppingCart';
 import { AuthContext } from '../../contexts/AuthContext';
@@ -13,9 +22,6 @@ const Wishlist = ({data, favourites, addItemToCart, isFaveFound, removeFromFavou
   useEffect(() => {
     const populateFavouriteItems = (favourites) => {
       if(!favourites.length) return []
-      console.log(!favourites.length)
-      // console.log(favouriteItems === favourites)
-      console.log(favourites)
       return favourites.map(favourite => {
         let newFav
         for(const category in data.products) {
@@ -29,7 +35,6 @@ const Wishlist = ({data, favourites, addItemToCart, isFaveFound, removeFromFavou
       
     }
     const newFavouriteItems = populateFavouriteItems(favourites)
-    console.log(newFavouriteItems)
     setFavouriteItems(newFavouriteItems)
   // eslint-disable-next-line react-hooks/exhaustive-deps
   }, [favourites, data])
@@ -39,16 +44,13 @@ const Wishlist = ({data, favourites, addItemToCart, isFaveFound, removeFromFavou
   }
 
   const handleClickCart = (id) => {
-    console.log('handleClickCart', id)
     addItemToCart(id)
     removeFromFavourites(id)
   }
 
   const handleClickFave = (id) => {
-    console.log('handleClickFave', id)
     if(!isFaveFound(id)) return
     removeFromFavourites(id)
-    // isFaveFound(id) ? removeFromFavourites(id) : addToFavourites(id)
   }
 
   return (
@@ -69,20 +71,16 @@ const Wishlist = ({data, favourites, addItemToCart, isFaveFound, removeFromFavou
                         <Grid item xs={6} sm={4} md={3} key={`${name}-${id}`}>
                             <Card sx={{ maxWidth: 345, height: 320, position: 'relative' }}>
                                 <CardContent>
-                                    {/* <Typography gutterBottom variant="h8" component="div" sx={{position:'absolute', top: 0, left:0, right:0, marginRight: 0, marginLeft: 'auto', zIndex: 10, background: 'rgba(255,255,255,0)'}}> */}
-                                    {/* <Stack direction="row" sx={{position:'absolute', top: 0, left:0, right:0, marginRight: 0, marginLeft: 'auto', zIndex: 10, background: 'rgba(255,255,255,0)'}}> */}
                                     <Stack direction="row-reverse" sx={{background: 'rgba(255,255,255,0)'}}>
                                         <IconButton onClick={() => handleClickCart(id)} aria-label={`add product id ${id} to cart`}>
                                             <AddShoppingCartIcon id={`cart-${id}`} role='button' aria-label={`add product id ${id} to cart and remove from wishlist`} />
                                         </IconButton>
                                         { isFaveFound(id) &&
                                           <IconButton onClick={() => handleClickFave(id)} aria-label={`add product id ${id} to wishlist`}>
-                                              {/* <FavoriteIcon id={`fav-${id}`} role='button' sx={{ color: 'rgb(255,20,147)'}} aria-label={`add product id ${id} to wishlist`}/> */}
                                               <FavoriteIcon id={`fav-${id}`} role='button' sx={{ color: isFaveFound(id) ? 'rgb(255,20,147)' : 'rgb(0,0,0, 0.35)'}} aria-label={`remove product id ${id} from wishlist`}/>
                                           </IconButton>
                                         }
                                     </Stack>
-                                    {/* </Typography> */}
                                 </CardContent>
                                 <CardActionArea onClick={(e) => handleClickNavigate(e, category, subcategory)}>
                                     <CardMedia
@@ -103,7 +101,6 @@ const Wishlist = ({data, favourites, addItemToCart, isFaveFound, removeFromFavou
                                     </Typography>
                                 </CardContent>
                             </Card>
-                            
                         </Grid>
                         )
                     })
