+++ src/contexts/AuthContext.js
@@ -1,6 +1,6 @@
 import { createContext, useEffect, useState } from "react";
 import { onAuthStateChanged } from "firebase/auth";
-import { auth, signUpUser, signInUser, signOutUser } from '../utils/firebase.utils'
+import { auth } from '../utils/firebase.utils'
 
 export const AuthContext = createContext({
     user: null,
@@ -12,12 +12,9 @@ export const AuthContext = createContext({
 export const AuthProvider = ({ children }) => {
     const [user, setUser] = useState(null)
     const [isSignedIn, setIsSignedIn] = useState(false)
-    console.log('authprovider user', user)
-    console.log('signed in?', isSignedIn)
     useEffect(() => {
         const unsubscribe = onAuthStateChanged(auth, (user) => {
             if (user) {
-
                 setUser(user)
                 setIsSignedIn(true)
             } else {
