+++ src/index.js
@@ -8,13 +8,13 @@ import reportWebVitals from './reportWebVitals';
 
 const root = ReactDOM.createRoot(document.getElementById('root'));
 root.render(
-  // <React.StrictMode>
-    <BrowserRouter>
-        <AuthProvider>
+  <React.StrictMode>
+    <AuthProvider>
+      <BrowserRouter>
           <App />
-        </AuthProvider>
-    </BrowserRouter>
-  // </React.StrictMode>
+      </BrowserRouter>
+    </AuthProvider>
+  </React.StrictMode>
 );
 
 // If you want to start measuring performance in your app, pass a function
