+++ src/utils/firebase.utils.js
@@ -7,17 +7,12 @@ import {
   signOut,
   signInWithPopup,
   GoogleAuthProvider,
-  setPersistence, 
-  browserSessionPersistence 
 } from 'firebase/auth'
 import { 
   getFirestore, 
-  addDoc, 
   getDoc,
   setDoc,
-  getDocs, 
   doc,
-  collection,
   Timestamp 
 } from 'firebase/firestore'
 // TODO: Add SDKs for Firebase products that you want to use
@@ -40,51 +35,11 @@ const db = getFirestore()
 const GoogleProvider = new GoogleAuthProvider()
 // Initialize Firebase Authentication and get a reference to the service
 export const auth = getAuth(app);
-console.log(auth)
 
-export const setUserSession = async (email, password) => {
-  try {
-    // Existing and future Auth states are now persisted in the current
-    // session only. Closing the window would clear any existing state even
-    // if a user forgets to sign out.
-    // ...
-    // New sign-in will be persisted with session persistence.
-    await setPersistence(auth, browserSessionPersistence)
-    return signInUser(email, password)
-  } catch(error) {
-    const { code, message } = error
-      return {error: `${code}: ${message}`}
-  }
-  // setPersistence(auth, browserSessionPersistence)
-  //   .then(() => {
-  //     // Existing and future Auth states are now persisted in the current
-  //     // session only. Closing the window would clear any existing state even
-  //     // if a user forgets to sign out.
-  //     // ...
-  //     // New sign-in will be persisted with session persistence.
-  //     return signInWithEmailAndPassword(email, password);
-  //   })
-  //   .catch((error) => {
-  //     // Handle Errors here.
-      
-  //   });
-}
-
-//set auth state persistence to session for sign up too
 export const signUpUser = async (email, password) => {
   try {
     const userCredential = await createUserWithEmailAndPassword(auth, email, password)
-    console.log(userCredential)
     const user = userCredential.user
-    // await addDoc(collection(db, 'users'), {
-    //   uid: user.uid,
-    //   firstName,
-    //   lastName,
-    //   displayName: user.displayName,
-    //   email: user.email,
-    //   creationTime: user.metadata.creationTime
-    // })
-    console.log(user)
     return user
   } catch(error) {
     const { code, message } = error
@@ -99,7 +54,6 @@ export const signInUser = async (email, password) => {
   try {
     const userCredential = await signInWithEmailAndPassword(auth, email, password)
     const user = userCredential.user
-    console.log(user)
     return user
   } catch(error) {
     const { code, message } = error
@@ -113,7 +67,6 @@ export const signInUserWithGoogle = async () => {
     // This gives you a Google Access Token. You can use it to access the Google API.
     const credential = GoogleAuthProvider.credentialFromResult(result);
     const token = credential.accessToken;
-    console.log('token', token)
     // The signed-in user info.
     const user = result.user;
     // IdP data available using getAdditionalUserInfo(result)
@@ -171,36 +124,12 @@ export const getUserInfo = async (user) => {
   if(!user) return
   const userDocRef = doc(db, 'users', user.uid)
   const userSnapshot = await getDoc(userDocRef)
-  console.log('getUserInfo userSnapshot', userSnapshot)
   const userData = userSnapshot.data()
   const { creationDate } = userData
-  console.log(userData.creationDate !== null)
-  console.log(creationDate.toDate()) 
   if(userData.creationDate !== null) {
-    // userData.creationDate = userData.creationDate.toDate().toString()
     userData.creationDate = convertTimestamp(userData.creationDate)
   } 
-  console.log(userData)
   return userData
-  // try {
-    
-    //fetch user data from Firestore
-    //construct return data object from retrieved user data
-    // const userProfile = {
-    //   firstName,
-    //   lastName,
-    //   phone,
-    //   addressLine1,
-    //   addressLine2,
-    //   cityTownVillage,
-    //   stateProvinceRegion,
-    //   postCode,
-    //   country,
-    // }
-  // } catch(error) {
-  //   const { code, message } = error
-  //     return {error: `${code}: ${message}`}
-  // }
 }
 
 export const updateUserInfo = async (user, userData = {}) => {
@@ -215,11 +144,12 @@ export const updateUserInfo = async (user, userData = {}) => {
      await setDoc(userDocRef, newUserData)
   } catch(error) {
     const { code, message } = error
-      return {error: `${code}: ${message}`}
+    return {error: `${code}: ${message}`}
   }
   return newUserData
 }
 
+//NOT FINISHED YET
 export const addNewUserTransaction = async (user, transactionObject) => {
   if(!user) return
   const userDocRef = doc(db, 'users', user.uid)
